Added ipython/Jupyter Notebooks for tests.
road_test loads the OpenDRIVE XML, modifies the merge lane (corner cases are not completely implemented, because it would take to long and we don't use it). It plots the Roads on a picture of the driving-map (from AADC). It also checks if the predecessors and successors are not too far away from the end or beginning of the road. After all it saves the RoadList object to a file.

load_roads_test loads the pickle file, and does the same successor/predecessor test.

README holds some information, like in this commits.

map_only.png is used in the road_test notebook. Sorry for committing a Blob, Mario.